+++ vpnservice/bgpmanager/bgpmanager-impl/src/main/java/org/opendaylight/netvirt/bgpmanager/ConfigureBgpCli.java
@@ -168,7 +168,7 @@ public class ConfigureBgpCli extends OsgiCommandSupport {
         }
         for (Neighbors nbr : nbrs) {
           if (nbrIp.equals(nbr.getAddress().getValue())) {
-             return nbr.getRemoteAs().longValue();
+             return nbr.getRemoteAs();
           }
         }
         return -1;
